pyproject.toml
[project]
name = "learn-fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.11",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.24",
]
